+++ src/osm_easy_api/__init__.py
@@ -1,6 +1,6 @@
-VERSION = "0.2.0"
+VERSION = "0.3.0"
 
-from .data_classes import Node, Way, Relation, OsmChange, Action, Tags
+from .data_classes import Node, Way, Relation, OsmChange, Action, Tags, Changeset
 from .diff import Diff, Frequency
 
 from .api import Api